Return extra info when API errors are encountered
One problem with the current code is that if you issue a command and
it fails (something like xunknownMethod, xsliceIDNotFound or
xInvalidparam) there's no feedback to the caller to let them analyze
that info and make decisions.

This change adds a custom error type to capture and provide error
response details from the SolidFire endpoint.  This will make the
code, message and name avaialble to the caller so they can choose
to do things like wait/retry etc.

While we're at it, clean up some logging a big making it a bit
standardized (at least in the dispatcher) and provide some extra
logging when using debug level.