Don't use things like `log.Fatal` in packages
Using Fatal, or Panic it annoying for consumers of packages because
it can kill their service and in most cases they probably don't want
that.  Especially for calling something like an unimplemented call.

Instead just log or return an appropriate error and let the caller
decide what to do, but don't kill their app.

Addresses Issue: #2